added the WriteHandler code to the Java bindings
git-svn-id: http://svn.code.sf.net/p/nitro-nitf/code/trunk@354 7dc3ea37-dbd3-4c13-bb9a-7ccd57e75167

+++ java/nitf/src/java/nitf/ImageSegment.java
@@ -50,12 +50,12 @@ public final class ImageSegment extends NITFObject
      * 
      * @return An integer offset
      */
-    public native int getImageOffset();
+    public native long getImageOffset();
 
     /**
      * Retrieve the ending offset for the image
      * 
      * @return An integer offset
      */
-    public native int getImageEnd();
+    public native long getImageEnd();
 }
